# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: VirtuProfs
basePath: /api
host: localhost:5100
schemes:
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  passport:
    type: apiKey
    in: header
    name: Authorization
  admin:
    type: apiKey
    in: header
    name: Authorization
  teacher:
    type: apiKey
    in: header
    name: Authorization
paths:
...
  /callbacks:
    x-swagger-router-controller: CallbackController
    get:
      operationId: getCallbacks
      summary: Get all user callbacks
      tags:
        - callback
      security:
        - admin: []
        - teacher: []
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
  /callbacks/access:
    x-swagger-router-controller: CallbackController
    get:
      operationId: callbackAccess
      summary: Get callback Access
      tags:
        - callback
      security:
        - passport: []
        - teacher: []
      parameters:
        - name: subject
          in: query
          type: string
          description: subject
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
  /callback/decline:
    x-swagger-router-controller: CallbackController
    put:
      operationId: declineCallback
      tags:
        - callback
      security:
        - passport: []
      parameters:
        - in: body
          name: callback
          description: callback
          schema:
            type: object
            required:
               - subject
            properties:
              subject:
                type: string
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
  /callback:
    x-swagger-router-controller: CallbackController
    post:
      operationId: setCallback
      tags:
        - callback
      security:
        - passport: []
        - teacher: []
      parameters:
        - in: body
          name: callback
          description: set callback
          schema:
            type: object
            # required:
            #   - reason
            #   - calltime
            #   - calldate
            #   - subject
            properties:
              _id:
                type: string
              reason:
                type: string
              calltime:
                type: string
              calldate:
                type: string
              subject:
                type: string
              status:
                type: string
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
  /callback/status:
    x-swagger-router-controller: CallbackController
    get:
      operationId: userCallbackAccess
      summary: Get user Callback Access
      tags:
        - callback
      security:
        - passport: []
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
...
definitions:
  Response:
    type: object
    properties:
      success:
        type: boolean
      errorCode:
        type: string
      errorStack:
        type: array
        items:
            type: object
      message:
        type: string
      data:
        type: object
    required:
      - success
...